fix errors

+++ Detectors/TPC/workflow/src/ZSSpec.cxx
@@ -310,7 +310,7 @@ DataProcessorSpec getZStoDigitsSpec(std::vector<int> const& inputIds)
       for (int i = 0; i < NSectors; i++) {
         LOG(INFO) << "digits in sector " << i << " : " << outDigits[i].size();
         o2::tpc::TPCSectorHeader sh{i};
-        pc.outputs().snapshot(Output{gDataOriginTPC, "DIGITS", i, Lifetime::Timeframe, sh}, outDigits[i]);
+        pc.outputs().snapshot(Output{gDataOriginTPC, "DIGITS", (unsigned int)i, Lifetime::Timeframe, sh}, outDigits[i]);
       }
     };
     return processingFct;
